Add support for stateless backlight
This ideally is a of temporary solution.
In the future I want to turn the backlight on at every change mimicking the remote control behavior which fixes the UI going out of sync with the state (because after each input the backlight will always be on) but that needs more infrastructure changes.